+++ TexTransCore/ITTStorageBuffer.cs
@@ -1,7 +1,0 @@
-#nullable enable
-namespace net.rs64.TexTransCore
-{
-    public interface ITTStorageBuffer : ITTObject
-    {
-    }
-}